To make this more bullet-proof
I updated to Ruby 2.2 and RoR 4.2 and I have started to appear error when processing "... / robokassa / paid":
NoMethodError in RobokassaController # paid
undefined method `downcase 'for nil: NilClass

In this case, something like @params["SignatureValue"] is not defined.
This converts everything to string to start with. nil.to_s is an empty string by default, so it's safe.